+++ docs/changelog.md
@@ -1,13 +1,20 @@
 # Changelog
 
-### v0.23 - "*vision*"
+### v0.24 - "*wit*"
 
-Requires libvips v8.8.1.
+Requires libvips v8.9.0.
 
-#### v0.23.5 - TBD
+#### v0.24.0 - TBD
+
+* Drop support for Node.js 8.
+  [#1910](https://github.com/lovell/sharp/issues/1910)
 
 * Ensure correct colour output for 16-bit, 2-channel PNG input with ICC profile.
   [#2013](https://github.com/lovell/sharp/issues/2013)
+
+### v0.23 - "*vision*"
+
+Requires libvips v8.8.1.
 
 #### v0.23.4 - 5<sup>th</sup> December 2019